stylist/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
sentence-transformers>=2.6.0
qdrant-client>=1.8.0
Pillow>=10.2.0
pandas>=2.2.0
numpy>=1.26.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
colorthief>=0.2.1
colormath>=3.0.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
httpx>=0.27.0
pydantic>=2.6.0
openai>=1.0.0
